Main Menu Settings Menu & DeleteAllSaves Fix
We've fixed the Settings Menu in the Main Menu Scene, so it now displays properly, and have fixed the DeleteAllSaves option in the build so that it only deletes .json files rather than trying (And failing) to delete .log files.

+++ GPW - Space Station/Assets/Code/Scripts/Saving/JsonDataService.cs
@@ -83,7 +83,10 @@ namespace JSONSerialisation
         {
             foreach(string filePath in Directory.GetFiles(DATA_PATH))
             {
-                File.Delete(filePath);
+                if (Path.GetExtension(filePath).EndsWith(FILE_EXTENSION))
+                {
+                    File.Delete(filePath);
+                }
             }
         }
